backend/requirements.txt
aiofiles
fastapi[standard]
sqlmodel
pydantic
PyJWT
argon2-cffi
pydantic_settings
requests
Pillow
